requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic>=2.10.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.3.0
redis==5.0.8
cryptography==43.0.1
stripe==10.12.0
requests==2.32.3
bleach==6.1.0
httpx==0.27.2
aiohttp>=3.13.0
aiofiles>=23.2.0
pyinstaller>=6.11.0
beautifulsoup4>=4.12.0
